Add documentation. Add error checking for file existing in save()

+++ src/baseball/PlayerSerializer.java
@@ -3,17 +3,32 @@ package baseball;
 import java.io.*;
 import java.util.*;
 
+
+/**
+ * Helper class that creates a single serialized data file for an ArrayList of
+ * Players. Call save to replace the file with the list you provide, and load
+ * to retrieve that list.
+ * @author Andres Chorro
+ */
 public class PlayerSerializer {
 	private final String filePath;
 
+	/**
+	 * Initialize a PlayerSerializer that will store the list of players in a
+	 * given file.
+	 * @param filePath The file to store the players in
+	 */
 	public PlayerSerializer(String filePath) {
 		this.filePath = filePath;
 	}
 
 	public void save(List<Player> players) {
 		try {
+			// delete the file if it exists
 			File fileOut = new File(filePath);
-			fileOut.delete();
+			if (fileOut.exists() && !fileOut.isDirectory())
+				fileOut.delete();
+			
 			FileOutputStream outFile = new FileOutputStream(filePath);
 			ObjectOutputStream out = new ObjectOutputStream(outFile);
 			out.writeObject(players);
@@ -25,7 +40,7 @@ public class PlayerSerializer {
 		}
 	}
 	
-	public List<Player> load() {
+	public ArrayList<Player> load() {
 		try {
 	         FileInputStream fileIn = new FileInputStream(filePath);
 	         ObjectInputStream in = new ObjectInputStream(fileIn);
@@ -51,6 +66,7 @@ public class PlayerSerializer {
 		ps.save(players);
 		players.add(new Player("Erik", "Jennings", 23));
 		ps.save(players);
+		System.out.println(ps.load());
 		System.out.println(ps.load());
 	}
 }
